Version 1.3.6
25.07.2007 - Fix an error that caused mon_ctrl to crash when it called
             another program.
26.07.2007 - When using server= and the name specified is only an alias,
             process sf_smtp would try to connect to this alias. Fixed.
08.08.2007 - For the dialog where you can select event actions, added a
             button to toggle all events.
             directory, don't set this to the full value specified by
             the user (max files) at once. Allocate it in hunks until we
             reach the maximum specified by the user.
             (via redirecting the output), write to a file and then print
             the file.
12.08.2007 - Close open file descriptors for dialog show_elog.
13.08.2007 - If host alias has the error/warnings set offline or to
             acknowledge, the reason can be seen by clicking on the
             host alias field in afd_crl dialog.
30.08.2007 - AFDD now properly checks current AFD message list.
31.08.2007 - When checking if another AFD is active make the timeout
             tuneable via command line argument.
             DIR_CONFIG.
02.09.2007 - Added splice() support in copy_file() and to sf_loc to
             reduce system load.
04.09.2007 - Remove a race condition from error queue that could lead
             to a host with an error queue always set.
             it from the error queue.
             the error queue set.
09.09.2007 - Speedup function archive_file(), by removing some unnecessary
             system calls and/or replacing them by more efficient calls.
10.09.2007 - When bursting a job of the same job ID we must recheck if
             error queue flag is set. Otherwise it can happen that we
             service a job that is in error queue but do not unset it after
             being done.
             disabled AFD. Fixed.
11.09.2007 - When host name is specified dialog show_ilog responce to user
             input was very slow. Fixed.
12.09.2007 - Program jid_view now shows the directory name. It is now
             possible to supply multiple job ID's as parameters.
             list is no longer in order. Fixed.
13.09.2007 - When multiple DIR_CONFIG's are defined in AFD_CONFIG, show
             the full path of the DIR_CONFIG to which the job belongs
             when viewing the configuration for a host.
20.09.2007 - Added script afd_unused_infos to show, delete or move unused
             INFO-* files.
21.09.2007 - When directory option 'ignore size' or 'ignore file time' is
             set and the option 'delete unknown files' is set, it will
             delete those files if the size or time do not match.
             xsend_file dialog so it was not visible. Fixed.
             Fixed.
             reported in RECEIVE_LOG. This can be disabled with the
             configure option --disable-report_empty_dir_scan.
29.09.2007 - Implemented persistend connection for transmitting files.
             zombies. By inserting a zombie queue FD can now spawn
             much faster new process.
             stopped or disabled host that this is not possible.
10.10.2007 - Fetching files via HTTP together with the directory option
             'do not get dir list' did not work with file filters having
             %t modifier and other modifiers. Fixed.
03.11.2007 - Added option to use scheme file if the hostname matches with
             the nodname on which AFD is running.
08.11.2007 - Added configuration file 'local_interface.list' where one
             can configure local interfaceses so AFD can use scheme file
             on those interfaces.
10.11.2007 - Dialog afd_info shows in the protocol line if it is active,
             passive or extended active/passive mode.
16.11.2007 - When changing the name of the host alias, it could happen
             that no more files get distributed for the new host alias
             name.
23.11.2007 - Process dir_check will now continue to try scan a directory
             once it was not accessible.

+++ src/amg/check_files.c
@@ -152,7 +152,7 @@ extern struct fileretrieve_status *fra;
 extern struct delete_log          dl;
 #endif
 
-/* Local function prototype */
+/* Local function prototypes. */
 static int                        get_last_char(char *, off_t);
 #ifdef _POSIX_SAVED_IDS
 static int                        check_sgids(gid_t);
@@ -274,6 +274,8 @@ check_files(struct directory_entry *p_de,
             fra[p_de->fra_pos].dir_flag |= DIR_ERROR_SET;
             SET_DIR_STATUS(fra[p_de->fra_pos].dir_flag,
                            fra[p_de->fra_pos].dir_status);
+            error_action(p_de->alias, "start", DIR_ERROR_ACTION);
+            event_log(0L, EC_DIR, ET_EXT, EA_ERROR_START, "%s", p_de->alias);
          }
          unlock_region(fra_fd,
 #ifdef LOCK_DEBUG
@@ -807,7 +809,14 @@ check_files(struct directory_entry *p_de,
                               }
                               else
                               {
-                                 max_file_buffer = fra[p_de->fra_pos].max_copied_files;
+                                 if ((max_file_buffer + FILE_BUFFER_STEP_SIZE) >= fra[p_de->fra_pos].max_copied_files)
+                                 {
+                                    max_file_buffer = fra[p_de->fra_pos].max_copied_files;
+                                 }
+                                 else
+                                 {
+                                    max_file_buffer += FILE_BUFFER_STEP_SIZE;
+                                 }
                               }
                               REALLOC_RT_ARRAY(file_name_pool, max_file_buffer,
                                                MAX_FILENAME_LENGTH, char);
@@ -931,6 +940,52 @@ check_files(struct directory_entry *p_de,
                      }
                   }
 #endif /* WITH_DUP_CHECK */
+               }
+            }
+            else
+            {
+               if ((fra[p_de->fra_pos].delete_files_flag & UNKNOWN_FILES) &&
+                   ((fra[p_de->fra_pos].ignore_size != 0) ||
+                    ((fra[p_de->fra_pos].ignore_file_time != 0) &&
+                     ((fra[p_de->fra_pos].gt_lt_sign & IFTIME_GREATER_THEN) ||
+                      (fra[p_de->fra_pos].gt_lt_sign & IFTIME_EQUAL)))) &&
+                   ((current_time - stat_buf.st_mtime) > fra[p_de->fra_pos].unknown_file_time))
+               {
+                  if (unlink(fullname) == -1)
+                  {
+                     system_log(WARN_SIGN, __FILE__, __LINE__,
+                                "Failed to unlink() %s : %s",
+                                fullname, strerror(errno));
+                  }
+                  else
+                  {
+#ifdef _DELETE_LOG
+                     size_t dl_real_size;
+
+                     (void)strcpy(dl.file_name, p_dir->d_name);
+                     (void)sprintf(dl.host_name, "%-*s %x",
+                                   MAX_HOSTNAME_LENGTH, "-",
+                                   DEL_UNKNOWN_FILE);
+                     *dl.file_size = stat_buf.st_size;
+                     *dl.job_number = p_de->dir_id;
+                     *dl.file_name_length = strlen(p_dir->d_name);
+                     dl_real_size = *dl.file_name_length + dl.size +
+                                    sprintf((dl.file_name + *dl.file_name_length + 1),
+# if SIZEOF_TIME_T == 4
+                                            "dir_check() >%ld",
+# else
+                                            "dir_check() >%lld",
+# endif
+                                            (pri_time_t)diff_time);
+                     if (write(dl.fd, dl.data, dl_real_size) != dl_real_size)
+                     {
+                        system_log(ERROR_SIGN, __FILE__, __LINE__,
+                                   "write() error : %s", strerror(errno));
+                     }
+#endif /* _DELETE_LOG */
+                     files_in_dir--;
+                     bytes_in_dir -= stat_buf.st_size;
+                  }
                }
             }
          } /* if (S_ISREG(stat_buf.st_mode)) */
@@ -1211,7 +1266,14 @@ check_files(struct directory_entry *p_de,
                                        }
                                        else
                                        {
-                                          max_file_buffer = fra[p_de->fra_pos].max_copied_files;
+                                          if ((max_file_buffer + FILE_BUFFER_STEP_SIZE) >= fra[p_de->fra_pos].max_copied_files)
+                                          {
+                                             max_file_buffer = fra[p_de->fra_pos].max_copied_files;
+                                          }
+                                          else
+                                          {
+                                             max_file_buffer += FILE_BUFFER_STEP_SIZE;
+                                          }
                                        }
                                        REALLOC_RT_ARRAY(file_name_pool, max_file_buffer,
                                                         MAX_FILENAME_LENGTH, char);
@@ -1371,7 +1433,8 @@ check_files(struct directory_entry *p_de,
                       (fra[p_de->fra_pos].delete_files_flag & UNKNOWN_FILES))
                   {
                      diff_time = current_time - stat_buf.st_mtime;
-                     if (diff_time > fra[p_de->fra_pos].unknown_file_time)
+                     if ((diff_time > fra[p_de->fra_pos].unknown_file_time) &&
+                         (diff_time > DEFAULT_TRANSFER_TIMEOUT))
                      {
                         if (unlink(fullname) == -1)
                         {
@@ -1409,6 +1472,52 @@ check_files(struct directory_entry *p_de,
                            bytes_in_dir -= stat_buf.st_size;
                         }
                      }
+                  }
+               }
+            }
+            else
+            {
+               if ((fra[p_de->fra_pos].delete_files_flag & UNKNOWN_FILES) &&
+                   ((fra[p_de->fra_pos].ignore_size != 0) ||
+                    ((fra[p_de->fra_pos].ignore_file_time != 0) &&
+                     ((fra[p_de->fra_pos].gt_lt_sign & IFTIME_GREATER_THEN) ||
+                      (fra[p_de->fra_pos].gt_lt_sign & IFTIME_EQUAL)))) &&
+                   ((current_time - stat_buf.st_mtime) > fra[p_de->fra_pos].unknown_file_time))
+               {
+                  if (unlink(fullname) == -1)
+                  {
+                     system_log(WARN_SIGN, __FILE__, __LINE__,
+                                "Failed to unlink() %s : %s",
+                                fullname, strerror(errno));
+                  }
+                  else
+                  {
+#ifdef _DELETE_LOG
+                     size_t dl_real_size;
+
+                     (void)strcpy(dl.file_name, p_dir->d_name);
+                     (void)sprintf(dl.host_name, "%-*s %x",
+                                   MAX_HOSTNAME_LENGTH, "-",
+                                   DEL_UNKNOWN_FILE);
+                     *dl.file_size = stat_buf.st_size;
+                     *dl.job_number = p_de->dir_id;
+                     *dl.file_name_length = strlen(p_dir->d_name);
+                     dl_real_size = *dl.file_name_length + dl.size +
+                                    sprintf((dl.file_name + *dl.file_name_length + 1),
+# if SIZEOF_TIME_T == 4
+                                            "dir_check() >%ld",
+# else
+                                            "dir_check() >%lld",
+# endif
+                                            (pri_time_t)diff_time);
+                     if (write(dl.fd, dl.data, dl_real_size) != dl_real_size)
+                     {
+                        system_log(ERROR_SIGN, __FILE__, __LINE__,
+                                   "write() error : %s", strerror(errno));
+                     }
+#endif /* _DELETE_LOG */
+                     files_in_dir--;
+                     bytes_in_dir -= stat_buf.st_size;
                   }
                }
             }
@@ -1518,9 +1627,10 @@ done:
          fra[p_de->fra_pos].last_retrieval = current_time;
          if (fra[p_de->fra_pos].dir_flag & WARN_TIME_REACHED)
          {
-            fra[p_de->fra_pos].dir_flag ^= WARN_TIME_REACHED;
+            fra[p_de->fra_pos].dir_flag &= ~WARN_TIME_REACHED;
             SET_DIR_STATUS(fra[p_de->fra_pos].dir_flag,
                            fra[p_de->fra_pos].dir_status);
+            error_action(p_de->alias, "stop", DIR_WARN_ACTION);
          }
          receive_log(INFO_SIGN, NULL, 0, current_time,
 #if SIZEOF_OFF_T == 4
@@ -1535,6 +1645,16 @@ done:
          ABS_REDUCE_QUEUE(p_de->fra_pos, files_copied, *total_file_size);
       }
    }
+#ifdef REPORT_EMPTY_DIR_SCANS
+   else
+   {
+      if ((count_files == YES) || (count_files == PAUSED_REMOTE))
+      {
+         receive_log(INFO_SIGN, NULL, 0, current_time,
+                     "Received 0 files with 0 Bytes.");
+      }
+   }
+#endif
 #ifdef _WITH_PTHREAD
    if ((ret = pthread_mutex_unlock(&fsa_mutex)) != 0)
    {
@@ -1555,9 +1675,11 @@ done:
       fra[p_de->fra_pos].error_counter = 0;
       if (fra[p_de->fra_pos].dir_flag & DIR_ERROR_SET)
       {
-         fra[p_de->fra_pos].dir_flag ^= DIR_ERROR_SET;
+         fra[p_de->fra_pos].dir_flag &= ~DIR_ERROR_SET;
          SET_DIR_STATUS(fra[p_de->fra_pos].dir_flag,
                         fra[p_de->fra_pos].dir_status);
+         error_action(p_de->alias, "stop", DIR_ERROR_ACTION);
+         event_log(0L, EC_DIR, ET_EXT, EA_ERROR_END, "%s", p_de->alias);
       }
       unlock_region(fra_fd,
 #ifdef LOCK_DEBUG
